fix: fixed routes, controllers; several refactoring changes
