+++ app.js
@@ -1,67 +1,35 @@
+require('dotenv').config();
 const express = require('express');
+const helmet = require('helmet');
 const mongoose = require('mongoose');
 const bodyParser = require('body-parser');
-const rateLimit = require('express-rate-limit');
 const { errors } = require('celebrate');
-const { celebrate, Joi } = require('celebrate');
 const cors = require('cors');
 const cookieParser = require('cookie-parser');
-require('dotenv').config();
-const moviesRouter = require('./routes/movies.js');
-const usersRouter = require('./routes/users.js');
-const auth = require('./middlewares/auth.js');
-const { login, createUser } = require('./controllers/users.js');
+const limiter = require('./middlewares/limiter');
+const errorHandler = require('./middlewares/errorHandler');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
-const NotFoundError = require('./errors/NotFoundError');
+const router = require('./routes/index');
 
-const { PORT = 3000 } = process.env;
+const { PORT = 3000, MONGO = 'mongodb://localhost:27017/diplomadb' } = process.env;
+
 const app = express();
-const limiter = rateLimit({
-  windowMs: 15 * 60 * 1000,
-  max: 100
-});
 
+app.use(requestLogger);
 app.use('/', cors());
+app.use(helmet());
 app.use(cookieParser());
 app.use(limiter);
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
-app.use(requestLogger);
 
-app.post('/signup', celebrate({
-  body: Joi.object().keys({
-    email: Joi.string().required().email(),
-    password: Joi.string().trim().required().min(6),
-    name: Joi.string().required().min(2).max(30)
-  })
-}), createUser);
-
-app.post('/signin', celebrate({
-  body: Joi.object().keys({
-    email: Joi.string().required().email(),
-    password: Joi.string().required().min(6),
-    name: Joi.string().required().min(2).max(30)
-  })
-}), login);
-
-app.use(auth);
-
-app.use('/', auth, usersRouter);
-app.use('/', auth, moviesRouter);
-
-app.all('/*', (req, res, next) => {
-  throw new NotFoundError('Запрашиваемый ресурс не найден');
-});
-
+app.use('/', router);
 app.use(errorLogger);
 app.use(errors());
 
-app.use((err, req, res, next) => {
-  const { statusCode = 500, message } = err;
-  res.status(statusCode).send({ message: statusCode === 500 ? 'На сервере произошла ошибка' : message });
-});
+app.use(errorHandler);
 
-mongoose.connect('mongodb://localhost:27017/diplomadb', {
+mongoose.connect(MONGO, {
   useNewUrlParser: true,
   useCreateIndex: true,
   useFindAndModify: false,
