+++ controllers/users.js
@@ -64,7 +64,7 @@ module.exports.login = (req, res, next) => {
         httpOnly: true,
         sameSite: true
       })
-        .send({ message: 'Авторизация прошла успешно' });
+        .send({ token });
     })
     .catch(next);
 };
