+++ package.json
@@ -28,6 +28,7 @@
     "express": "^4.17.1",
     "express-rate-limit": "^5.2.5",
     "express-winston": "^4.0.5",
+    "helmet": "^4.4.1",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^5.11.15",
     "validator": "^13.5.2",
